web-animations/animation-timeline/animation-timeline-play-playback-rate.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test playbackRate attribute of AnimationPlayer created by AnimationTimeline.play()</title>
<meta name="assert" content="AnimationTimeline.play() creates AnimationPlayer object with playbackRate attribute initially set to 1">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationTimeline-play-AnimationPlayer-AnimationNode-source">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var player = document.timeline.play(newAnimation(createDiv(this)));

    assert_equals(player.playbackRate, 1,
        'Initial value of the playbackRate attribute of the newly-created AnimationPlayer ' +
        'object should be 1');
}, 'Test AnimationTimeline.play() creates AnimationPlayer object with playbackRate ' +
    'attribute initially set to 1');

test(function() {
    var player = document.timeline.play(null);

    assert_equals(player.playbackRate, 1,
        'Initial value of the playbackRate attribute of the newly-created AnimationPlayer ' +
        'object should be 1');
}, 'Test AnimationTimeline.play(null) creates AnimationPlayer object with playbackRate ' +
    'attribute initially set to 1');

test(function() {
    var player = document.timeline.play();

    assert_equals(player.playbackRate, 1,
        'Initial value of the playbackRate attribute of the newly-created AnimationPlayer ' +
        'object should be 1');
}, 'Test AnimationTimeline.play() without arguments creates AnimationPlayer object ' +
    'with playbackRate attribute initially set to 1');
</script>
</body>
